requirements.txt
Flask
gunicorn
pandas
numpy
-e .
PyYAML
PyYAML
os